Clean code for prune-tune and upgrade readme
